Mistake in parsing nested paraentheses in expr3() function.
In some instances returned value before completion of the parsing.
In other instances assigned value to temp variable but failed to
return the variable carrying the result of the epxression.

Fixed all found instances so I am not potentially confused in the
future if I come back to reference some of this code.