Completed Lesson 17.
Implemented call stack and activation records. This is
an improvement in the interpreter global memory system.
It allows storage of procedure and function calls, as well,
as variables corresponding to those calls. It also stores
program, or global, level variables.

It will allow for the development of storing nested procedure and
function calls, and recursive function calls.