Completed Lesson 19.
Been a long time coming. I have finally completed AND feel
that I have a decent understanding of the lessons up to this
final published one. I hope there are more.

In Lesson 19 I implemented nested procedure calls successfully.

I had to go back and re-read the prior 7 or 8 lessons because
I got really confused about what the semantic analyzer was doing,
and how the references were working to get access from the
ProcDecl block AST node and formal parameter list in the ProcDecl
Symbol to the ProcCall AST node. At least in the explanation given
here I thought the purpose of the semantic analyzer is not well
defined.It needs a well defined list of items it is checking off and
where in the code that is being handled and why. I ended up writing
one out for myself. Maybe I should include that as a comment in the code.
It appeared to be a very subtle explanation. Reading the first chapter or so
of Parr's Language Implementation Patterns I decided that it was
important to understand more deeply what the semantic analyzer
is doing. Armed with a little more context from Parr I re-read
many lessons here to have a better understanding. All the information
is there but like I said, the function of the semantic analyzer is
not directly addressed well, it is too subtle.